Спринт 12 (уже не успел в дедлайн)

+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserDbStorage.java
@@ -91,11 +91,10 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public void removeFriend(Long userId, Long friendId) {
-        jdbcTemplate.update(
-                "DELETE FROM friends WHERE user_id = ? AND friend_id = ?",
-                userId, friendId
-        );
+        jdbcTemplate.update("DELETE FROM friends WHERE user_id = ? AND friend_id = ?", userId, friendId);
+        jdbcTemplate.update("DELETE FROM friends WHERE user_id = ? AND friend_id = ?", friendId, userId);
     }
+
 
     @Override
     public List<User> getFriends(Long userId) {
